src/main/java/com/example/cloudDiplom/service/UserService.java
package com.example.cloudDiplom.service;


import com.example.cloudDiplom.entity.User;
import com.example.cloudDiplom.repository.UsersRepo;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class UserService implements UserDetailsService {
    private final UsersRepo usersRepo;

    @Override
    public UserDetails loadUserByUserName(String userName) throws UsernameNotFoundException {
        User user = usersRepo.findByUserName(userName);

        if (user == null) {
            throw new UsernameNotFoundException("Пользователь не найден");
        }
        return user;
    }
}